VMS-10351: Fixed conditional logging to multiple destinations
<==== Mercurial data ====>
Branch: x_VMS-10351_logs

+++ common_libs/nx_utils/src/nx/utils/log/log_settings.cpp
@@ -78,7 +78,7 @@ void Settings::load(const QnSettings& settings, const QString& prefix)
     }
 
     // If there are more prefix/* arguments, adding one more logger.
-    if (loggers.size() < logSettingCount)
+    if ((int) loggers.size() < logSettingCount)
         loadCompatibilityLogger(settings, prefix);
 }
